Support the new `extended tweet_mode` API (#200)
Twitter's API has changed and now shortens statuses that are longer
than 140 chars, but have an acceptable length by the new Twitter
rules.

This change introduces Tweet.FullText, which contains the
unshortened form of the status. Tweet.Text's behavior should be
unchanged and will contain the shortened form. All API requests are
now executed in the new extended mode.

Sadly the streaming API now behaves slightly different to the regular
REST API and contains the extended Tweet information in a separate
struct. This change also parses this extra struct and correctly merges
the status into the Tweet struct, so both API modes behave the same
and remain backwards compatible, as far as anaconda users are concerned.

This PR makes #158 and #190 obsolete.